+++ src/Codec/Archive/Pack.hs
@@ -5,36 +5,38 @@ module Codec.Archive.Pack ( entriesToFile
                           , entriesToBSzip
                           , entriesToBS7zip
                           , packEntries
+                          , noFail
                           ) where
 
 import           Codec.Archive.Foreign
+import           Codec.Archive.Monad
 import           Codec.Archive.Types
-import           Control.Monad         (void)
-import           Data.ByteString       (packCStringLen, useAsCStringLen)
-import qualified Data.ByteString       as BS
-import           Data.Foldable         (traverse_)
-import           Data.Semigroup        (Sum (..))
+import           Control.Monad          (void)
+import           Control.Monad.IO.Class (MonadIO (..))
+import           Data.ByteString        (packCStringLen)
+import qualified Data.ByteString        as BS
+import           Data.Foldable          (traverse_)
+import           Data.Semigroup         (Sum (..))
 import           Foreign.C.String
-import           Foreign.Marshal.Alloc (alloca, allocaBytes)
-import           Foreign.Ptr           (Ptr)
-import           Foreign.Storable      (peek)
-import           System.IO.Unsafe      (unsafePerformIO)
+import           Foreign.Ptr            (Ptr)
+import           Foreign.Storable       (peek)
+import           System.IO.Unsafe       (unsafePerformIO)
 
-contentAdd :: EntryContent -> Ptr Archive -> Ptr ArchiveEntry -> IO ()
+contentAdd :: EntryContent -> Ptr Archive -> Ptr ArchiveEntry -> ArchiveM ()
 contentAdd (NormalFile contents) a entry = do
-    archive_entry_set_filetype entry regular
-    archive_entry_set_size entry (fromIntegral (BS.length contents))
-    void $ archive_write_header a entry
-    useAsCStringLen contents $ \(buff, sz) ->
-        void $ archive_write_data a buff (fromIntegral sz)
+    liftIO $ archive_entry_set_filetype entry regular
+    liftIO $ archive_entry_set_size entry (fromIntegral (BS.length contents))
+    handle $ archiveWriteHeader a entry
+    useAsCStringLenArchiveM contents $ \(buff, sz) ->
+        liftIO $ void $ archive_write_data a buff (fromIntegral sz)
 contentAdd Directory a entry = do
-    archive_entry_set_filetype entry directory
-    void $ archive_write_header a entry
+    liftIO $ archive_entry_set_filetype entry directory
+    handle $ archiveWriteHeader a entry
 contentAdd (Symlink fp) a entry = do
-    archive_entry_set_filetype entry symlink
-    withCString fp $ \fpc ->
+    liftIO $ archive_entry_set_filetype entry symlink
+    liftIO $ withCString fp $ \fpc ->
         archive_entry_set_symlink entry fpc
-    void $ archive_write_header a entry
+    handle $ archiveWriteHeader a entry
 
 setOwnership :: Ownership -> Ptr ArchiveEntry -> IO ()
 setOwnership (Ownership uname gname uid gid) entry =
@@ -50,7 +52,7 @@ setOwnership (Ownership uname gname uid gid) entry =
 setTime :: ModTime -> Ptr ArchiveEntry -> IO ()
 setTime (time', nsec) entry = archive_entry_set_mtime entry time' nsec
 
-packEntries :: (Foldable t) => Ptr Archive -> t Entry -> IO ()
+packEntries :: (Foldable t) => Ptr Archive -> t Entry -> ArchiveM ()
 packEntries a = traverse_ (archiveEntryAdd a)
 
 -- Get a number of bytes appropriate for creating the archive.
@@ -65,36 +67,45 @@ entriesSz = getSum . foldMap (Sum . entrySz)
 --
 -- @since 1.0.0.0
 entriesToBS :: Foldable t => t Entry -> BS.ByteString
-entriesToBS = unsafePerformIO . entriesToBSGeneral archive_write_set_format_pax_restricted
+entriesToBS = unsafePerformIO . noFail . entriesToBSGeneral archive_write_set_format_pax_restricted
 {-# NOINLINE entriesToBS #-}
 
 -- | Returns a 'BS.ByteString' containing a @.7z@ archive with the 'Entry's
 --
 -- @since 1.0.0.0
 entriesToBS7zip :: Foldable t => t Entry -> BS.ByteString
-entriesToBS7zip = unsafePerformIO . entriesToBSGeneral archive_write_set_format_7zip
+entriesToBS7zip = unsafePerformIO . noFail . entriesToBSGeneral archive_write_set_format_7zip
 {-# NOINLINE entriesToBS7zip #-}
 
 -- | Returns a 'BS.ByteString' containing a zip archive with the 'Entry's
 --
 -- @since 1.0.0.0
 entriesToBSzip :: Foldable t => t Entry -> BS.ByteString
-entriesToBSzip = unsafePerformIO . entriesToBSGeneral archive_write_set_format_zip
+entriesToBSzip = unsafePerformIO . noFail . entriesToBSGeneral archive_write_set_format_zip
 {-# NOINLINE entriesToBSzip #-}
 
+-- This is for things we don't think will fail. When making a 'BS.ByteString'
+-- from a bunch of 'Entry's, for instance, we don't anticipate any errors
+noFail :: ArchiveM a -> IO a
+noFail act = do
+    res <- runArchiveM act
+    case res of
+        Right x -> pure x
+        Left _  -> error "Should not fail."
+
 -- | Internal function to be used with 'archive_write_set_format_pax' etc.
-entriesToBSGeneral :: (Foldable t) => (Ptr Archive -> IO ArchiveError) -> t Entry -> IO BS.ByteString
+entriesToBSGeneral :: (Foldable t) => (Ptr Archive -> IO ArchiveError) -> t Entry -> ArchiveM BS.ByteString
 entriesToBSGeneral modifier hsEntries' = do
-    a <- archive_write_new
-    void $ modifier a
-    alloca $ \used ->
-        allocaBytes bufSize $ \buffer -> do
-            void $ archive_write_open_memory a buffer bufSize used
+    a <- liftIO archive_write_new
+    ignore $ modifier a
+    allocaArchiveM $ \used ->
+        allocaBytesArchiveM bufSize $ \buffer -> do
+            handle $ archiveWriteOpenMemory a buffer bufSize used
             packEntries a hsEntries'
-            void $ archive_write_close a
-            usedSz <- peek used
-            res <- curry packCStringLen buffer (fromIntegral usedSz)
-            void $ archive_write_free a
+            handle $ archiveWriteClose a
+            usedSz <- liftIO $ peek used
+            res <- liftIO $ curry packCStringLen buffer (fromIntegral usedSz)
+            ignore $ archive_write_free a
             pure res
 
     where bufSize :: Integral a => a
@@ -108,44 +119,44 @@ entriesToBSGeneral modifier hsEntries' = do
 -- @
 --
 -- @since 1.0.0.0
-entriesToFile :: Foldable t => FilePath -> t Entry -> IO ()
+entriesToFile :: Foldable t => FilePath -> t Entry -> ArchiveM ()
 entriesToFile = entriesToFileGeneral archive_write_set_format_pax_restricted
 -- this is the recommended format; it is a tar archive
 
 -- | Write some entries to a file, creating a zip archive.
 --
 -- @since 1.0.0.0
-entriesToFileZip :: Foldable t => FilePath -> t Entry -> IO ()
+entriesToFileZip :: Foldable t => FilePath -> t Entry -> ArchiveM ()
 entriesToFileZip = entriesToFileGeneral archive_write_set_format_zip
 
 -- | Write some entries to a file, creating a @.7z@ archive.
 --
 -- @since 1.0.0.0
-entriesToFile7Zip :: Foldable t => FilePath -> t Entry -> IO ()
+entriesToFile7Zip :: Foldable t => FilePath -> t Entry -> ArchiveM ()
 entriesToFile7Zip = entriesToFileGeneral archive_write_set_format_7zip
 
-entriesToFileGeneral :: Foldable t => (Ptr Archive -> IO ArchiveError) -> FilePath -> t Entry -> IO ()
+entriesToFileGeneral :: Foldable t => (Ptr Archive -> IO ArchiveError) -> FilePath -> t Entry -> ArchiveM ()
 entriesToFileGeneral modifier fp hsEntries' = do
-    a <- archive_write_new
-    void $ modifier a
-    withCString fp $ \fpc ->
-        void $ archive_write_open_filename a fpc
+    a <- liftIO archive_write_new
+    ignore $ modifier a
+    withCStringArchiveM fp $ \fpc ->
+        handle $ archiveWriteOpenFilename a fpc
     packEntries a hsEntries'
-    void $ archive_write_free a
+    ignore $ archive_write_free a
 
-withArchiveEntry :: (Ptr ArchiveEntry -> IO a) -> IO a
+withArchiveEntry :: MonadIO m => (Ptr ArchiveEntry -> m a) -> m a
 withArchiveEntry fact = do
-    entry <- archive_entry_new
+    entry <- liftIO archive_entry_new
     res <- fact entry
-    archive_entry_free entry
+    liftIO $ archive_entry_free entry
     pure res
 
-archiveEntryAdd :: Ptr Archive -> Entry -> IO ()
+archiveEntryAdd :: Ptr Archive -> Entry -> ArchiveM ()
 archiveEntryAdd a (Entry fp contents perms owner mtime) =
     withArchiveEntry $ \entry -> do
-        withCString fp $ \fpc ->
+        liftIO $ withCString fp $ \fpc ->
             archive_entry_set_pathname entry fpc
-        archive_entry_set_perm entry perms
-        setOwnership owner entry
-        setTime mtime entry
+        liftIO $ archive_entry_set_perm entry perms
+        liftIO $ setOwnership owner entry
+        liftIO $ setTime mtime entry
         contentAdd contents a entry